requirements.txt
gradio[mcp]
python-dotenv
textblob
smolagents[mcp]
langchain-mcp-adapters
langgraph
langchain-google-genai